Rename io.rs -> fastq.rs
