+++ extract/src/lib.rs
@@ -1,6 +1,6 @@
 pub mod barcode;
 pub mod error;
-pub mod io;
+pub mod fastq;
 pub mod logger;
 pub mod pattern;
 pub mod run;
