+++ extract/src/logger.rs
@@ -2,7 +2,7 @@ use console::Emoji;
 use indicatif::{ProgressBar, ProgressStyle};
 
 use crate::error::Error;
-use crate::io;
+use crate::fastq;
 
 pub static SPARKLE: Emoji<'_, '_> = Emoji("✨ ", ":-)");
 
@@ -16,7 +16,7 @@ pub fn create_progress_bar(
     ).expect("Failed to parse a progress bar template")
     .progress_chars("##-");
 
-    let read_num = io::get_reads_count(fastq_file, threads, max_memory) as u64;
+    let read_num = fastq::get_reads_count(fastq_file, threads, max_memory) as u64;
 
     let progress_bar = ProgressBar::new(read_num as u64);
     progress_bar.set_style(progress_bar_style.clone());
